Profile Edit into Component
Continuing to break up functionality into separate files
