+++ frontend-next/src/app/user/[stub]/page.js
@@ -3,18 +3,17 @@
 import { useState, useEffect } from "react";
 import { auth, database, storage } from "../../../../firebase-config";
 import { ref, onValue, get, update } from "firebase/database";
-import { ref as sRef, getDownloadURL } from "firebase/storage";
 import { onAuthStateChanged } from "firebase/auth";
-import { useForm, Form } from "react-hook-form";
 
 // Refactored Component Imports
+
 // Data Structure Imports
 import { ProfileRoom } from "../../../components/app/profile/ProfileRoom";
+import { ProfileEdit } from "../../../components/app/profile/ProfileEdit";
 import { Interest } from "../../../components/app/profile/Interest";
 
 // Header Import
 import { Header } from "../../../components/app/header";
-import { uploadBytes } from "firebase/storage";
 
 // Contains most everything for the app homepage
 function Home({ params }) {
@@ -26,9 +25,13 @@ function Home({ params }) {
   const [userInterestArray, setUserInterestArray] = useState(null);
   const [userRoomsArray, setUserRoomsArray] = useState(null);
   const [isOwner, setIsOwner] = useState(false);
+
   const [isEditing, setIsEditing] = useState(false);
 
-  var { register, control } = useForm();
+  //Handles Edit State in Component, shares useState with ProfileEdit
+  const handleIsEditing = (newValue) => {
+    setIsEditing(newValue);
+  };
 
   // Authentication
   useEffect(() => {
@@ -58,6 +61,10 @@ function Home({ params }) {
     onValue(ref(database, "/users/" + params.stub), (snapshot) => {
       setProfileData(snapshot.val());
       var interests = snapshot.val().interests || null;
+      if (interests == null) {
+        // Placeholder for no interests specified, will be replaced with default interests
+        interests = "Music, Sports, Movies";
+      }
       interests = interests.split(",");
       var interestArray = [];
       var i = 0;
@@ -75,36 +82,6 @@ function Home({ params }) {
       setUserRoomsArray(roomArray);
     });
   }, []);
-
-  function save({ data }) {
-    if (data.pfp[0]) {
-      // image stuff
-      uploadBytes(sRef(storage, `users/${user.uid}/pfp`), data.pfp[0]).then(
-        () => {
-          getDownloadURL(sRef(storage, `users/${user.uid}/pfp`)).then((url) => {
-            data.pfp = url;
-            for (var key in data) {
-              if (data[key] == "") {
-                data[key] = profileData[key];
-              }
-            }
-
-            setIsEditing(false);
-            update(ref(database, `users/${user.uid}`), data);
-          });
-        }
-      );
-    } else {
-      for (var key in data) {
-        if (data[key] == "") {
-          data[key] = profileData[key];
-        }
-      }
-      data.pfp = profileData.pfp;
-      setIsEditing(false);
-      update(ref(database, `users/${user.uid}`), data);
-    }
-  }
 
   return (
     <div>
@@ -154,90 +131,7 @@ function Home({ params }) {
                   </div>
                 )}
                 {isEditing && (
-                  <div>
-                    <Form
-                      onSubmit={save}
-                      encType={"application/json"}
-                      control={control}
-                    >
-                      <div className="grid grid-cols-2">
-                        <div>
-                          <img
-                            src={profileData.pfp}
-                            width="150px"
-                            className="relative mx-auto rounded-2xl overflow-hidden"
-                          />
-                          Current Profile Picture
-                        </div>
-                        <div className="flex content-center">
-                          <input
-                            type="file"
-                            {...register("pfp")}
-                            className="w-[80%]"
-                            accept=".jpg,.png,.jpeg"
-                          />
-                        </div>
-                      </div>
-                      <div className="grid grid-cols-2 pl-2 w-[90%]">
-                        <div className="pt-5">
-                          <div className="font-bold">First Name</div>
-                          <input
-                            className="w-[80%] border-2 border-gray-300 p-2 rounded-lg"
-                            type="text"
-                            {...register("firstName")}
-                            placeholder={profileData.firstName}
-                          />
-                        </div>
-                        <div className="pt-5">
-                          <div className="font-bold">Last Name</div>
-                          <input
-                            className="w-[80%] border-2 border-gray-300 p-2 rounded-lg"
-                            type="text"
-                            {...register("lastName")}
-                            placeholder={profileData.lastName}
-                          />
-                        </div>
-                        <div className="pt-5">
-                          <div className="font-bold">Username</div>
-                          <input
-                            className="w-[80%] border-2 border-gray-300 p-2 rounded-lg"
-                            type="text"
-                            {...register("username")}
-                            placeholder={profileData.username}
-                          />
-                        </div>
-                        <div className="pt-5">
-                          <div className="font-bold">
-                            Interests (Comma Seperated)
-                          </div>
-                          <input
-                            className="w-[80%] border-2 border-gray-300 p-2 rounded-lg"
-                            type="text"
-                            {...register("interests")}
-                            placeholder={profileData.interests}
-                          />
-                        </div>
-                        <div className="pt-5 col-span-2">
-                          <div className="font-bold">Bio</div>
-                          <textarea
-                            className="w-[92%] border-2 border-gray-300 p-2 rounded-lg"
-                            {...register("bio")}
-                            type="text"
-                            placeholder={profileData.bio}
-                          />
-                        </div>
-                        <div className="justify-items-center pt-5 col-span-2">
-                          <button
-                            type="submit"
-                            className="p-2 cursor-pointer bg-cyan-500 text-white font-bold rounded-full text-center"
-                          >
-                            {" "}
-                            Save Changes{" "}
-                          </button>
-                        </div>
-                      </div>
-                    </Form>
-                  </div>
+                  <ProfileEdit profileData={profileData} user={user} onSave={handleIsEditing}/>
                 )}
               </div>
               <div className="col-span-2">
