+++ frontend-next/src/components/app/map/geo.js
@@ -2,6 +2,7 @@ import { Map, Marker, ZoomControl } from "pigeon-maps";
 
 // Map module for main page and chat room sidebar (and eventually user profile)
 // Constructs Map and Markers
+// TODO: Need to get rest of marker handling here or in marker file.
 export function Geo({ loc, zoom, locMarker, markers }) {
   if (loc) {
     return (
